Added x axis line to the bar chart.
Added a separate line to the x-axis of the bar chart so that the axis
would appear to span the width of the chart. Otherwise the generic axis
would go from the middle of the leftmost bar to the middle of the
rightmost bar. (Styling for this may need to be added in the future.)