Line Graph can now handle multiple lines.
The Line Graph is now capable of displaying multiple line if given a
DataColumns list with more than two column indices. Each line will be
displayed with a different random color, and the highlighting color
matches. There is now a method in visualizer.js that returns a random
RGB value as a string (i.e. "rgb(125,37,245)").