Added guideline to Line graph.
A guideline now follows the mouse cursor in the Line Graph.
Still to do: (1) Make it only display when within the axes. (2) Make it
still display even when cursor hovers over another element, like the
points. (3) Make it so that when the line is aligned with points, they
highlight and the hover popup displays.