Redesigned multiline line graph text highlights.
I redesigned and reimplemented the text popups that occur when the user
selects a column of points with the guideline in the multiline line
graph. This method avoids the pop-up overlap problem that occurred when
each data point had its own popup horzontally-aligned with it. The Line
class code is in major need of commenting and refactoring, but I'll get
to that.