Pie and Treemap can now graph 1-col tables.
If the getData() function receives instructions to extract one column
from the table, it will give the visualization two copies of that
column. For example, if the instructions say DataColumns[0], getData
will pull 0 twice. When handled by the Treemap and Pie, this results in
a proportional mapping to the pie/tree. It's possible this could pose an
issue if further vis types are allowed to only have 1 column, but it
works for our purposes.